# Lab2/ark-pzpi-23-3-osychenko-illia-lab2/docs/swagger.yaml
basePath: /api
definitions:
...
  handler.SyncEventsRequest:
    properties:
      events:
        items:
          properties:
            event_type:
              type: string
            latitude:
              type: number
            local_id:
              type: integer
            longitude:
              type: number
            passenger_count_after:
              type: integer
            timestamp:
              type: string
          type: object
        type: array
      trip_id:
        type: integer
    type: object
...
  model.Bus:
    properties:
      capacity:
        example: 50
        type: integer
      fuel_consumption_per_100km:
        example: 12.5
        type: number
      id:
        example: 1
        type: integer
      is_active:
        example: true
        type: boolean
      model:
        example: Mercedes Sprinter
        type: string
      registration_number:
        example: AA1234BB
        type: string
    type: object
...
  model.Route:
    properties:
      base_price:
        example: 250
        type: number
      created_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      destination_city:
        example: Київ
        type: string
      distance_km:
        example: 480.5
        type: number
      driver_cost_per_trip:
        example: 800
        type: number
      estimated_duration_minutes:
        example: 360
        type: integer
      fuel_cost_per_km:
        example: 2.5
        type: number
      id:
        example: 1
        type: integer
      is_active:
        example: true
        type: boolean
      origin_city:
        example: Харків
        type: string
      updated_at:
        example: "2023-01-01T00:00:00Z"
        type: string
    type: object
...
  service.SyncEventsResponse:
    properties:
      last_synced_local_id:
        type: integer
      server_time:
        type: string
      synced_count:
        type: integer
      trip_current_passengers:
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    email: support@busoptima.ua
    name: API Support
  description: API для системи оптимізації автобусних перевезень BusOptima
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: BusOptima API
  version: "1.0"
paths:
...
  /buses:
    get:
      consumes:
      - application/json
      description: Повертає список всіх автобусів з можливістю фільтрації
      parameters:
      - description: Тільки активні автобуси
        in: query
        name: active_only
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Bus'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Отримати список автобусів
      tags:
      - Buses
    post:
      consumes:
      - application/json
      description: Створює новий автобус в системі
      parameters:
      - description: Дані автобуса
        in: body
        name: bus
        required: true
        schema:
          $ref: '#/definitions/model.Bus'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Bus'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Створити новий автобус
      tags:
      - Buses
  /buses/{id}:
    delete:
      consumes:
      - application/json
      description: Видаляє автобус за ID
      parameters:
      - description: ID автобуса
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Видалити автобус
      tags:
      - Buses
    get:
      consumes:
      - application/json
      description: Повертає автобус за вказаним ідентифікатором
      parameters:
      - description: ID автобуса
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Bus'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Отримати автобус за ID
      tags:
      - Buses
    put:
      consumes:
      - application/json
      description: Оновлює існуючий автобус за ID
      parameters:
      - description: ID автобуса
        in: path
        name: id
        required: true
        type: integer
      - description: Оновлені дані автобуса
        in: body
        name: bus
        required: true
        schema:
          $ref: '#/definitions/model.Bus'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Bus'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Оновити автобус
      tags:
      - Buses
  /iot/config/{tripId}:
    get:
      consumes:
      - application/json
      description: Повертає конфігурацію рейсу для IoT-пристрою
      parameters:
      - description: ID рейсу
        in: path
        name: tripId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Отримати конфігурацію рейсу
      tags:
      - IoT
  /iot/events:
    post:
      consumes:
      - application/json
      description: Отримує та зберігає події входу/виходу пасажирів від IoT-пристрою
      parameters:
      - description: Події пасажирів
        in: body
        name: events
        required: true
        schema:
          $ref: '#/definitions/handler.SyncEventsRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/service.SyncEventsResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Синхронізація подій пасажирів
      tags:
      - IoT
  /iot/price:
    post:
      consumes:
      - application/json
      description: Отримує рекомендацію ціни від IoT-пристрою
      parameters:
      - description: Рекомендація ціни
        in: body
        name: recommendation
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Отримати рекомендацію ціни
      tags:
      - IoT
  /routes:
    get:
      consumes:
      - application/json
      description: Повертає список всіх маршрутів з можливістю фільтрації
      parameters:
      - description: Тільки активні маршрути
        in: query
        name: active_only
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Route'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Отримати список маршрутів
      tags:
      - Routes
    post:
      consumes:
      - application/json
      description: Створює новий маршрут в системі
      parameters:
      - description: Дані маршруту
        in: body
        name: route
        required: true
        schema:
          $ref: '#/definitions/model.Route'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Route'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Створити новий маршрут
      tags:
      - Routes
  /routes/{id}:
    delete:
      consumes:
      - application/json
      description: Видаляє маршрут за ID
      parameters:
      - description: ID маршруту
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Видалити маршрут
      tags:
      - Routes
    get:
      consumes:
      - application/json
      description: Повертає маршрут за вказаним ідентифікатором
      parameters:
      - description: ID маршруту
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Route'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Отримати маршрут за ID
      tags:
      - Routes
    put:
      consumes:
      - application/json
      description: Оновлює існуючий маршрут за ID
      parameters:
      - description: ID маршруту
        in: path
        name: id
        required: true
        type: integer
      - description: Оновлені дані маршруту
        in: body
        name: route
        required: true
        schema:
          $ref: '#/definitions/model.Route'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Route'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Оновити маршрут
      tags:
      - Routes
...
securityDefinitions:
  BearerAuth:
    description: 'JWT Authorization header using the Bearer scheme. Example: "Authorization:
      Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"